+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "typia",
-  "version": "4.0.6",
+  "version": "4.0.8",
   "description": "Superfast runtime validators with only one line",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
