+++ src/factories/internal/metadata/iterate_metadata_resolve.ts
@@ -7,21 +7,32 @@ import { Writable } from "../../../typings/Writable";
 import { MetadataCollection } from "../../MetadataCollection";
 import { MetadataFactory } from "../../MetadataFactory";
 import { TypeFactory } from "../../TypeFactory";
-import { explore_metadata } from "./explore_metadata";
+import { iterate_metadata } from "./iterate_metadata";
+
+// import { iterate_metadata_coalesce } from "./iterate_metadata_coalesce";
 
 export const iterate_metadata_resolve =
     (checker: ts.TypeChecker) =>
     (options: MetadataFactory.IOptions) =>
     (collection: MetadataCollection) =>
-    (meta: Metadata, type: ts.Type, parentResolved: boolean): boolean => {
-        if (options.resolve === true && parentResolved === false) {
-            const resolved: ts.Type | null = TypeFactory.resolve(checker)(type);
-            if (resolved !== null) {
-                Writable(meta).resolved = explore_metadata(checker)(options)(
-                    collection,
-                )(resolved, true);
-                return true;
-            }
-        }
-        return false;
+    (
+        meta: Metadata,
+        type: ts.Type,
+        resolved: boolean,
+        aliased: boolean,
+    ): boolean => {
+        if (options.resolve === false || resolved === true) return false;
+
+        const escaped: ts.Type | null = TypeFactory.resolve(checker)(type);
+        if (escaped === null) return false;
+
+        if (meta.resolved === null)
+            Writable(meta).resolved = Metadata.initialize();
+        iterate_metadata(checker)(options)(collection)(
+            meta.resolved!,
+            escaped,
+            true,
+            aliased,
+        );
+        return true;
     };
