+++ test/schemas/json/swagger/ToJsonUnion.json
@@ -10,7 +10,19 @@
         "type": "array",
         "items": {
           "oneOf": [
-            {},
+            {
+              "oneOf": [
+                {
+                  "type": "boolean"
+                },
+                {
+                  "$ref": "#/components/schemas/ToJsonUnion.ICitizen"
+                },
+                {
+                  "$ref": "#/components/schemas/ToJsonUnion.IProduct"
+                }
+              ]
+            },
             {
               "type": "string"
             },
@@ -18,13 +30,7 @@
               "type": "number"
             },
             {
-              "type": "boolean"
-            },
-            {
               "$ref": "#/components/schemas/ToJsonUnion.ICitizen"
-            },
-            {
-              "$ref": "#/components/schemas/ToJsonUnion.IProduct"
             }
           ]
         }
